% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{berlin_fr_roof}
\alias{berlin_fr_roof}
\title{Heights distribution of buildings in Berlin}
\format{An array with dimensions "rlon", "rlat", "uheight1", "udir" and
"uclass" (see \link{ugrid}). In this example, only one urban class is considered.}
\source{
From \href{http://doi.org/10.1127/0941-2948/2013/0393}{Schubert and
  Grossman-Clarke 2013}.
}
\usage{
berlin_fr_roof
}
\description{
Height distribution of buildings in Berlin. The grid is defined in
\link{berlin_grid}.
}
\keyword{datasets}
