% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{berlin_fr_uclass}
\alias{berlin_fr_uclass}
\title{Fraction of urban classes in Berlin}
\format{An array with dimensions "rlon", "rlat"  and "uclass" (see
  \link{ugrid}). In this example, only one urban class is considered.}
\source{
From \href{http://doi.org/10.1127/0941-2948/2013/0393}{Schubert and
  Grossman-Clarke 2013}.
}
\usage{
berlin_fr_uclass
}
\description{
Fraction of different urban classes in Berlin. The grid is defined in
\link{berlin_grid}. Since only one urban class is considered in this example,
the value is 1 where urban values are defined.
}
\keyword{datasets}
