% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{berlin_fr_urb}
\alias{berlin_fr_urb}
\title{Urban fraction in Berlin}
\format{An array with dimensions "rlon" and "rlat" (see \link{ugrid}).}
\source{
From \href{http://doi.org/10.1127/0941-2948/2013/0393}{Schubert and
  Grossman-Clarke 2013}.
}
\usage{
berlin_fr_urb
}
\description{
Urban fraction of a grid cell in Berlin. The grid is defined in
\link{berlin_grid}.
}
\keyword{datasets}
