% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{berlin_grid}
\alias{berlin_grid}
\title{Used grid for the parameters of Berlin}
\format{Object of class \code{ugrid} (see \link{ugrid}).}
\source{
From \href{http://doi.org/10.1127/0941-2948/2013/0393}{Schubert and
  Grossman-Clarke 2013}.
}
\usage{
berlin_grid
}
\description{
Urban grid used with urban Berlin files.
}
\keyword{datasets}
