% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcz_default}
\alias{lcz_default}
\title{Default LCZ parameter}
\format{Object of class \code{lcz}.}
\usage{
lcz_default
}
\description{
Default Local Climate Zones parameter.
}
\keyword{datasets}
