% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf_output.R
\name{nc_def_var}
\alias{nc_def_var}
\title{Define NetCDF variable with standard attributes}
\usage{
nc_def_var(nc, varname, dimensions, standard_name, long_name, units)
}
\arguments{
\item{nc}{NetCDF file object.}

\item{varname}{String of variable name.}

\item{dimensions}{Vector of dimension names as strings.}

\item{standard_name}{String of standard name.}

\item{long_name}{String of long name.}

\item{units}{String of units.}
}
\description{
The variable is defined and attributes used in COSMO-CLM are added.
}
\examples{
nc <- RNetCDF::create.nc("rot_lon.nc")
RNetCDF::dim.def.nc(nc, dimname = "rlon", dimlength = 100)
dcepucp:::nc_def_var(nc, varname = "rlon", dimensions = "rlon",
                     standard_name = "grid_longitude",
                     long_name = "rotated longitude",
                     units = "degrees")
RNetCDF::var.put.nc(nc, variable ="rlon", data = 1:100)
RNetCDF::close.nc(nc)
}
\keyword{internal}
