% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz.R
\name{print.lcz}
\alias{print.lcz}
\title{Show LCZ class parameter, ranges and validate settings}
\usage{
\method{print}{lcz}(x, ...)
}
\arguments{
\item{x}{LCZ parameter of class "lcz".}

\item{...}{Not used.}
}
\description{
Print the Local Climate Zones parameter, its ranges and validate it.
}
\examples{
print(lcz_default)
}
