% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{print.ugrid}
\alias{print.ugrid}
\title{Print urban grid information}
\usage{
\method{print}{ugrid}(x, ...)
}
\arguments{
\item{x}{Object of class \code{ugrid}.}

\item{...}{Other parameters not used here.}
}
\description{
Prints parameters of an urban grid.
}
\examples{
grid <- ugrid(pollat = 40., pollon = -70.6,
              startlat_tot = -0.882, startlon_tot = -0.882,
              ie_tot = 195, je_tot = 195,
              dlat = 0.009, dlon = 0.009,
              ke_uhl = 8, hhl_uhl = c(0, 5, 10, 15, 20, 25, 30, 35,40),
              n_uclass = 1,
              n_udir = 2, angle_udir = c(0., 90.))
# calls print.ugrid
print(grid)
}
