% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz.R
\name{set_lcz_parameter}
\alias{set_lcz_parameter}
\title{Set LCZ parameter}
\usage{
set_lcz_parameter(lcz, lcz_name, h2w = NULL, fr_build = NULL,
  fr_street = NULL, height = NULL)
}
\arguments{
\item{lcz}{LCZ parameter, object of class \code{lcz}.}

\item{lcz_name}{Number or character string of the LCZ.}

\item{h2w}{The building-height-to-street-width ratio.}

\item{fr_build}{Fraction of buildings.}

\item{fr_street}{Fraction of street surfaces.}

\item{height}{Average building height.}
}
\description{
Modify one or several parameters of one Local Climate Zones. The values are
checked for consistency with the valid LCZ range.
}
\examples{
lcz <- set_lcz_parameter(lcz_default, "Compact high-rise", h2w = 3)
lcz
}
