% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{upar}
\alias{upar}
\title{Define a set of urban parameters}
\usage{
upar(grid, fr_urb, fr_uclass, fr_udir, fr_roof, w_street, w_build)
}
\arguments{
\item{grid}{Object of class \code{ugrid}.}

\item{fr_urb}{Fraction of urban surfaces in a grid cell and, thus, defines
urban grid cells.}

\item{fr_uclass}{Fraction of each urban class in each urban grid cell. The
sum for each urban grid cell is 1.}

\item{fr_udir}{Fraction of each urban class and street direction in each grid
cell. The sum for each urban grid cell is 1.}

\item{fr_roof}{Height distribution of each urban class and street direction
in each grid cell.}

\item{w_street}{Street width of each urban class and street direction in each
grid cell.}

\item{w_build}{Building width of each urban class and street direction in
each grid cell.}
}
\value{
An object of class \code{upar}.
}
\description{
Function to define a set of urban parameters. This includes building and
street properties as well as the urban grid. Some consistency checks are
done.
}
\examples{
# use Berlin data included in dcepucp
ucps <- upar(berlin_grid, fr_urb = berlin_fr_urb,
             fr_uclass = berlin_fr_uclass, fr_udir = berlin_fr_udir,
             fr_roof = berlin_fr_roof, w_street = berlin_w_street,
             w_build = berlin_w_build)
}
