% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf_output.R
\name{upar2nc}
\alias{upar2nc}
\title{Save urban parameters into NetCDF}
\usage{
upar2nc(file, ucp)
}
\arguments{
\item{file}{String of output NetCDF file.}

\item{ucp}{\code{upar} object.}
}
\description{
An \code{upar} object is stored in a NetCDF file. This file includes all
variables of the input object with a standard set of NetCDF attributes.
Information about the rotated pole are included as well.
}
\examples{
ucps <- upar(berlin_grid, fr_urb = berlin_fr_urb,
             fr_uclass = berlin_fr_uclass, fr_udir = berlin_fr_udir,
             fr_roof = berlin_fr_roof, w_street = berlin_w_street,
             w_build = berlin_w_build)
upar2nc("ucp.nc", ucps)
}
