% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz.R
\name{validate_lcz}
\alias{validate_lcz}
\title{Validate a parameter set}
\usage{
validate_lcz(lcz)
}
\arguments{
\item{lcz}{The LCZ parameter set.}
}
\description{
A Local Climate Zones parameter set is validated against parameter ranges.
}
\examples{
validate_lcz(lcz_default)
}
