% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{print.upar}
\alias{print.upar}
\title{Print urban parameters information}
\usage{
\method{print}{upar}(ucp)
}
\arguments{
\item{ucp}{Object of class \code{upar}.}
}
\description{
Prints details of an \code{upar} object.
}
\examples{
# use Berlin data included in dcepucp
ucps <- upar(berlin_grid, fr_urb = berlin_fr_urb,
             fr_uclass = berlin_fr_uclass, fr_udir = berlin_fr_udir,
             fr_roof = berlin_fr_roof, w_street = berlin_w_street,
             w_build = berlin_w_build)
print.upar(ucps)
ucps                       # also calls print.ucps
}
