% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ugrid}
\alias{ugrid}
\title{Define an urban grid}
\usage{
ugrid(pollat, pollon, polgam = 0, dlat, dlon, startlat_tot, startlon_tot,
  ie_tot, je_tot, ke_uhl, hhl_uhl, n_uclass = 1, n_udir = 2,
  angle_udir = c(0, 90))
}
\arguments{
\item{pollat}{Geographical latitude of the rotated north pole (in degrees,
north >0); for a non-rotated lat-lon grid set \code{pollat = 90}.}

\item{pollon}{Geographical longitude of the rotated north pole (in degrees,
east >0); for a non-rotated lat-lon grid set \code{pollon = -180}.}

\item{polgam}{Angle between the north poles of two rotated grids (in degrees,
east > 0)}

\item{dlat}{'Meridional' (rotated lat-direction) grid spacing (in degrees).}

\item{dlon}{'Zonal' (rotated lon-direction) grid spacing (in degrees).}

\item{startlat_tot}{Latitude of the lower left scalar grid point of the total
domain (in degrees, north >0, rotated coordinates).}

\item{startlon_tot}{Longitude of the lower left scalar grid point of the
total domain (in degrees, east >0, rotated coordinates).}

\item{ie_tot}{Number of gridpoints of the total domain in 'west-east'
direction of the rotated coordinates.}

\item{je_tot}{Number of gridpoints of the total domain in 'south-north'
direction of the rotated coordinates.}

\item{ke_uhl}{Number of main urban height levels corresponding to the number
of wall elements.}

\item{hhl_uhl}{Vector of the height of urban half levels. The \code{ke_uhl+1}
values are the heights of the roof levels.}

\item{n_uclass}{Number of urban classes within a grid cell.}

\item{n_udir}{Number considered street orientations for each urban class.}

\item{angle_udir}{Vector of street orientations relative to the south-north
axis.}
}
\value{
An object of class \code{ugrid}.
}
\description{
Function to define an urban grid including COSMO-CLM grid parameters. Some
consistency checks are done.
}
\note{
In \link{upar2nc}, \code{dlon} and \code{startlon_tot} as well as
  \code{dlat} and \code{startlat_tot} are used to define the dimensions
  "rlon" and "rlat", respectively; \code{ke_uhl} and \code{hhl_uhl} for
  "uheight1"; \code{n_udir} and \code{angle_udir} for "udir"; \code{n_uclass}
  for "uclass".
}
\examples{
grid <- ugrid(pollat = 40., pollon = -70.6,
              startlat_tot = -0.882, startlon_tot = -0.882,
              ie_tot = 195, je_tot = 195,
              dlat = 0.009, dlon = 0.009,
              ke_uhl = 8, hhl_uhl = c(0, 5, 10, 15, 20, 25, 30, 35,40),
              n_uclass = 1,
              n_udir = 2, angle_udir = c(0., 90.))
}
