% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcz.R
\name{upar_lcz}
\alias{upar_lcz}
\title{Generate DCEP paramters from LCZ maps}
\usage{
upar_lcz(grid, lcz, maps, mode = "dominant_class")
}
\arguments{
\item{grid}{Object of class \code{ugrid}.}

\item{lcz}{Object of class \code{lcz}.}

\item{maps}{LCZ maps}

\item{mode}{Mode}
}
\value{
An object of class \code{upar}.
}
\description{
Generate DCEP paramters from LCZ maps
}
\examples{
berlin_grid <- ugrid(pollat = 37.483, pollon = -166.6,
startlat_tot = -0.882, startlon_tot = -0.882,
ie_tot = 4, je_tot = 5,
dlat = 0.009, dlon = 0.009,
ke_uhl = 3,
hhl_uhl = c(0, 5, 10, 15),
n_uclass = 1,
n_udir = 4, angle_udir = c(-45., 0., 45., 90.))

up <- upar_lcz(berlin_grid, lcz_default, maps = list("Open low-rise" = array(1, c(4,5))))
}
